+++ h2integrate/transporters/pipe.py
@@ -24,6 +24,7 @@ class PipePerformanceModel(om.ExplicitComponent):
                 "wellhead_gas",
                 "water",
                 "oxygen",
+                "heat",
             ],
         )
         self.options.declare("plant_config", types=dict)
@@ -41,6 +42,8 @@ class PipePerformanceModel(om.ExplicitComponent):
             units = "galUS/h"
         elif transport_item == "co2":
             units = "kg/h"
+        elif transport_item == "heat":
+            units = "kW"
         else:
             units = "kg/s"